+++ sdk/netapp/arm-netapp/src/operations/volumes.ts
@@ -2738,9 +2738,9 @@ const listNextOperationSpec: coreClient.OperationSpec = {
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
+    Parameters.nextLink,
     Parameters.resourceGroupName,
     Parameters.accountName,
-    Parameters.nextLink,
     Parameters.poolName,
   ],
   headerParameters: [Parameters.accept],