+++ sdk/netapp/arm-netapp/CHANGELOG.md
@@ -1,15 +1,94 @@
 # Release History
+    
+## 21.0.0 (2024-05-14)
+    
+**Features**
 
-## 20.0.1 (Unreleased)
+  - Added operation group BackupsUnderAccount
+  - Added operation group BackupsUnderBackupVault
+  - Added operation group BackupsUnderVolume
+  - Added operation group BackupVaults
+  - Added operation group NetAppResourceRegionInfos
+  - Added operation Backups.beginCreate
+  - Added operation Backups.beginCreateAndWait
+  - Added operation Backups.beginDelete
+  - Added operation Backups.beginDeleteAndWait
+  - Added operation Backups.beginUpdate
+  - Added operation Backups.beginUpdateAndWait
+  - Added operation Backups.get
+  - Added operation Backups.getLatestStatus
+  - Added operation Backups.getVolumeLatestRestoreStatus
+  - Added operation Backups.listByVault
+  - Added Interface Backup
+  - Added Interface BackupPatch
+  - Added Interface BackupRestoreFiles
+  - Added Interface BackupsCreateOptionalParams
+  - Added Interface BackupsDeleteHeaders
+  - Added Interface BackupsDeleteOptionalParams
+  - Added Interface BackupsGetLatestStatusOptionalParams
+  - Added Interface BackupsGetOptionalParams
+  - Added Interface BackupsGetVolumeLatestRestoreStatusOptionalParams
+  - Added Interface BackupsList
+  - Added Interface BackupsListByVaultNextOptionalParams
+  - Added Interface BackupsListByVaultOptionalParams
+  - Added Interface BackupsMigrationRequest
+  - Added Interface BackupStatus
+  - Added Interface BackupsUnderAccountMigrateBackupsHeaders
+  - Added Interface BackupsUnderAccountMigrateBackupsOptionalParams
+  - Added Interface BackupsUnderBackupVaultRestoreFilesHeaders
+  - Added Interface BackupsUnderBackupVaultRestoreFilesOptionalParams
+  - Added Interface BackupsUnderVolumeMigrateBackupsHeaders
+  - Added Interface BackupsUnderVolumeMigrateBackupsOptionalParams
+  - Added Interface BackupsUpdateHeaders
+  - Added Interface BackupsUpdateOptionalParams
+  - Added Interface BackupVault
+  - Added Interface BackupVaultPatch
+  - Added Interface BackupVaultsCreateOrUpdateOptionalParams
+  - Added Interface BackupVaultsDeleteHeaders
+  - Added Interface BackupVaultsDeleteOptionalParams
+  - Added Interface BackupVaultsGetOptionalParams
+  - Added Interface BackupVaultsList
+  - Added Interface BackupVaultsListByNetAppAccountNextOptionalParams
+  - Added Interface BackupVaultsListByNetAppAccountOptionalParams
+  - Added Interface BackupVaultsUpdateHeaders
+  - Added Interface BackupVaultsUpdateOptionalParams
+  - Added Interface NetAppResourceRegionInfosGetOptionalParams
+  - Added Interface NetAppResourceRegionInfosListNextOptionalParams
+  - Added Interface NetAppResourceRegionInfosListOptionalParams
+  - Added Interface RegionInfoResource
+  - Added Interface RegionInfosList
+  - Added Interface VolumeBackupProperties
+  - Added Type Alias BackupsCreateResponse
+  - Added Type Alias BackupsDeleteResponse
+  - Added Type Alias BackupsGetLatestStatusResponse
+  - Added Type Alias BackupsGetResponse
+  - Added Type Alias BackupsGetVolumeLatestRestoreStatusResponse
+  - Added Type Alias BackupsListByVaultNextResponse
+  - Added Type Alias BackupsListByVaultResponse
+  - Added Type Alias BackupsUnderAccountMigrateBackupsResponse
+  - Added Type Alias BackupsUnderBackupVaultRestoreFilesResponse
+  - Added Type Alias BackupsUnderVolumeMigrateBackupsResponse
+  - Added Type Alias BackupsUpdateResponse
+  - Added Type Alias BackupType
+  - Added Type Alias BackupVaultsCreateOrUpdateResponse
+  - Added Type Alias BackupVaultsDeleteResponse
+  - Added Type Alias BackupVaultsGetResponse
+  - Added Type Alias BackupVaultsListByNetAppAccountNextResponse
+  - Added Type Alias BackupVaultsListByNetAppAccountResponse
+  - Added Type Alias BackupVaultsUpdateResponse
+  - Added Type Alias NetAppResourceRegionInfosGetResponse
+  - Added Type Alias NetAppResourceRegionInfosListNextResponse
+  - Added Type Alias NetAppResourceRegionInfosListResponse
+  - Interface VolumeBackups has a new optional parameter volumeResourceId
+  - Interface VolumePatchPropertiesDataProtection has a new optional parameter backup
+  - Interface VolumePropertiesDataProtection has a new optional parameter backup
+  - Added Enum KnownBackupType
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Removed operation Backups.getVolumeRestoreStatus
+    
+    
 ## 20.0.0 (2024-03-05)
     
 **Features**